+++ src/components/index.ts
@@ -9,8 +9,10 @@ export { SnackbarProvider } from './common/SnackbarProvider'
 export { Board } from './views/Board'
 export { Home } from './views/Home'
 export { ArchivedBoardModal } from './board/ArchivedBoardModal'
-export { BoardDrawer } from './board/BoardDrawer'
 export { BoardListSearchForm } from './board/BoardListSearchForm'
 export { BoardTitle } from './board/BoardTitle'
 export { CreateBoardModal } from './board/CreateBoardModal'
 export { CardList } from './list/CardList'
+export { BoardDrawer } from './board/drawer/BoardDrawer'
+export { DrawerArchivedItem } from './board/drawer/DrawerArchivedItem'
+export { DrawerRoot } from './board/drawer/DrawerRoot'
